% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info_accidents}
\alias{info_accidents}
\title{Variable descriptions for get_accidents()}
\usage{
info_accidents(lang = "en")
}
\arguments{
\item{lang}{\code{"en"} (default) or \code{"pt"}.}
}
\value{
A tibble with columns \code{variable}, \code{type}, \code{description}.
}
\description{
Variable descriptions for get_accidents()
}
\examples{
info_accidents()
info_accidents(lang = "pt")
}
