% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info_crashes}
\alias{info_crashes}
\title{Variable descriptions for get_crashes()}
\usage{
info_crashes(lang = "en")
}
\arguments{
\item{lang}{\code{"en"} (default) or \code{"pt"}.}
}
\value{
A tibble with columns \code{variable}, \code{type}, \code{description}.
}
\description{
Variable descriptions for get_crashes()
}
