% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info_violations}
\alias{info_violations}
\title{Variable descriptions for get_violations()}
\usage{
info_violations(lang = "en")
}
\arguments{
\item{lang}{\code{"en"} (default) or \code{"pt"}.}
}
\value{
A tibble with columns \code{variable}, \code{type}, \code{description}.
}
\description{
Variable descriptions for get_violations()
}
\examples{
info_violations(lang = "pt")
}
