% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{prf_cache_clear}
\alias{prf_cache_clear}
\title{Delete locally cached PRF data files}
\usage{
prf_cache_clear(dataset = NULL, year = NULL)
}
\arguments{
\item{dataset}{Character or NULL. Dataset to clear. NULL = all.}

\item{year}{Integer or NULL. Year(s) to clear. NULL = all years.}
}
\value{
Invisible NULL.
}
\description{
Delete locally cached PRF data files
}
\examples{
\dontrun{
prf_cache_clear("violations", year = 2024)
prf_cache_clear()
}
}
