% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprf-package.R
\docType{package}
\name{tidyprf-package}
\alias{tidyprf}
\alias{tidyprf-package}
\title{tidyprf: Tidy Access to Brazilian Federal Highway Police (PRF) Data}
\description{
Provides tidy-formatted access to Brazilian Federal Highway Police (PRF) open data: traffic accidents by person, accidents by occurrence, and infractions. Data is downloaded from GitHub Releases on demand and cached locally for reuse.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
