% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprf-package.R
\docType{package}
\name{tidyprf-package}
\alias{tidyprf}
\alias{tidyprf-package}
\title{tidyprf: Tidy Access to Brazilian Federal Highway Police ('PRF') Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Download and read Brazilian Federal Highway Police ('Polícia Rodoviária Federal', 'PRF') open data on traffic accidents by person, by occurrence, and traffic violations \url{https://www.gov.br/prf/pt-br/acesso-a-informacao/dados-abertos/dados-abertos-da-prf}. Data are retrieved on demand as 'Parquet' files distributed via 'GitHub' releases and cached locally for reuse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bonijoao/tidyprf}
  \item \url{https://bonijoao.github.io/tidyprf/}
  \item Report bugs at \url{https://github.com/bonijoao/tidyprf/issues}
}

}
\author{
\strong{Maintainer}: João Paulo Assis Bonifácio \email{jpab.27@hotmail.com} (\href{https://orcid.org/0009-0009-2335-8391}{ORCID})

Authors:
\itemize{
  \item Geraldo Magela da Cruz Pereira (\href{https://orcid.org/0000-0001-6280-4870}{ORCID})
  \item Pedro Mambelli Fernandes
}

}
\keyword{internal}
